license_portal/requirements.txt
Django==3.2
psycopg2
celery==5.2.7
djangorestframework==3.14.0
django-celery-results==2.4.0
python-dateutil==2.8.2
freezegun==1.2.2
django-cors-headers==3.13.0